% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/function.Bayesian.R
\name{Bayesian}
\alias{Bayesian}
\title{The function to perform BSA analysis with empirical Bayesian algorithm}
\usage{
Bayesian(
  file,
  Bulksize,
  Genomelength,
  Windowsize,
  threshold = 0.95,
  pval = 0.05
)
}
\arguments{
\item{file}{The data frame exported by BSA_filter()}

\item{Bulksize}{The number of samples in each pool}

\item{Genomelength}{The total length of reference genome}

\item{Windowsize}{A numeric value indicates the window size (in base pairs) containing SNPs subjected to calculating empirical Bayesian}

\item{threshold}{A numeric threshold value. Default value is "0.95"}

\item{pval}{A numeric threshold value. Default value is "0.05"}
}
\value{
a list containing three data frame,the data frame named "total.result" containing all the calculate result by empirical Bayesian algorithm,the data frame named "sigthreshold.result" containing the significant SNPs under fractile quantile set by parameter "threshold" calculated by empirical Bayesian algorithm,the data frame named "sigpval.result" containing the significant SNPs selected by P-value set by parameter "pval" calculated by empirical Bayesian algorithm
}
\description{
The function to perform BSA analysis with empirical Bayesian algorithm
}
