% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/function.filter.R
\name{BSA_filter}
\alias{BSA_filter}
\title{The function to filter SNPs based on read depth and quality of each bulk}
\usage{
BSA_filter(
  file,
  Bulk.DP,
  Bulk2.Ref,
  Bulk2.Alt,
  Depth.diff,
  min.GQ,
  verbose = TRUE
)
}
\arguments{
\item{file}{The data frame exported by vcf2table() or the data frame who has the same structure as the output of vcf2table()}

\item{Bulk.DP}{The minimum total read depth for a SNP}

\item{Bulk2.Ref}{The minimum read depth of a reference SNP allele}

\item{Bulk2.Alt}{The minimum read depth of a alternative SNP allele}

\item{Depth.diff}{The maximum absolute differences between the read depths of the bulks}

\item{min.GQ}{The minimum Genotype Quality output by GATK}

\item{verbose}{Default value is “TRUE”. If the value is "TRUE", the number of SNPs filtered in each step will print}
}
\value{
a data frame containing SNPs after filtering
}
\description{
The function to filter SNPs based on read depth and quality of each bulk
}
