% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/function.DEG.R
\name{DEG_analysis}
\alias{DEG_analysis}
\title{The function to perform differential expression analysis}
\usage{
DEG_analysis(file, rep.num, sig.level, file.name, exp.fold, plot = TRUE)
}
\arguments{
\item{file}{A data frame containing read count of each pool}

\item{rep.num}{A numeric value indicates the number of replications in differential expression analysis}

\item{sig.level}{numeric value indicates the FDR threshold in differential expression analysis}

\item{file.name}{A character to name the output plot}

\item{exp.fold}{A numeric value indicates the log2FoldChange threshold in differential expression analysis}

\item{plot}{Default value is "TRUE". If "TRUE", the volcano plot will be created.}
}
\value{
a data frame of the differential expression analysis result and a volcano plot of analysis result if parameter "plot" is set "TRUE"
}
\description{
The function to perform differential expression analysis
}
