% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/function.DeltaSNP.R
\name{DeltaSNP}
\alias{DeltaSNP}
\title{The function to perform BSA analysis with △SNP algorithm}
\usage{
DeltaSNP(file, Windowsize, threshold, pval)
}
\arguments{
\item{file}{The data frame exported by BSA_filter()}

\item{Windowsize}{A numeric value indicates the window size (in base pairs) containing SNPs subjected to calculating △SNP}

\item{threshold}{From 0 to 1, default value is “0.95”. A numeric threshold value to extract candidate region by fractile quantile}

\item{pval}{A numeric threshold value, default value is “0.05”}
}
\value{
a list containing three data frame,the data frame named "total.result" containing all the calculate result by △SNP algorithm,the data frame named "sigthreshold.result" containing the significant SNPs under fractile quantile set by parameter "threshold" calculated by △SNP algorithm,the data frame named "sigpval.result" containing the significant SNPs selected by P-value set by parameter "pval" calculated by G△SNP algorithmm
}
\description{
The function to perform BSA analysis with △SNP algorithm
}
