% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/function.ED.R
\name{ED}
\alias{ED}
\title{The function to perform BSA analysis with Euclidean distance algorithm}
\usage{
ED(file, power = 4, Windowsize, threshold = 0.95, pval = 0.05)
}
\arguments{
\item{file}{The data frame exported by BSA_filter()}

\item{power}{A numeric value indicates the power to raise ED. It was suggested to be 4~6,default value is "4"}

\item{Windowsize}{A numeric value indicates the window size (in base pairs) containing SNPs subjected to calculating Euclidean distance}

\item{threshold}{From 0 to 1, default value is "0.95". A numeric threshold value to extract candidate region by fractile quantile}

\item{pval}{A numeric threshold value, default value is "0.05"}
}
\value{
a list containing three data frame,the data frame named "total.result" containing all the calculate result by Euclidean Distance algorithm,the data frame named "sigthreshold.result" containing the significant SNPs under fractile quantile set by parameter "threshold" calculated by Euclidean Distance algorithm,the data frame named "sigpval.result" containing the significant SNPs selected by P-value set by parameter "pval" calculated by Euclidean Distance algorithm
}
\description{
The function to perform BSA analysis with Euclidean distance algorithm
}
