% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/function.GO.R
\name{GO_analysis}
\alias{GO_analysis}
\title{The function to perform GO analysis}
\usage{
GO_analysis(
  GO_term,
  DEG_id,
  type = c("BP", "MF", "CC", "all"),
  algorithm = c("classic", "weight01", "elim"),
  statistic = c("fisher", "ks"),
  top_num,
  plot = TRUE
)
}
\arguments{
\item{GO_term}{A data frame contains at least two columns, one column is genes' name and the other column is the GO term}

\item{DEG_id}{A vector containing the names of the differential expression genes}

\item{type}{A character indicates the mode of GO analysis. There are four options: "CC" indicates cellular component; "BP" indicates biological process, "MF" indicates molecular function, and "all" indicates all three modes. The default is “all”}

\item{algorithm}{A character indicates the algorithm for GO analysis. Three options "classic","weight01", and "elim" indicate three different algorithms}

\item{statistic}{A character indicates two different significance test methods. Two options "fisher" and "ks" indicates two different methods}

\item{top_num}{A numeric value to indicate the number of significant GO terms}

\item{plot}{Default value is “TRUE”. If "TRUE", the acyclic, bubble, and histogram plots will be created}
}
\value{
a data frame of the gene ontology analysis result and plot directed acyclic graphs, bubble graphs, histograms if parameter plot is set "TRUE"
}
\description{
The function to perform GO analysis
}
