% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GO_Term.example.R
\docType{data}
\name{GO_Term.example}
\alias{GO_Term.example}
\title{GO Terms example data}
\format{
Dataframe, one row per gene and a go term, with two columns
\describe{
\item{Gene id}{Gene symbol/name for gene ontology analysis refers to. Data type: character}
\item{GO Term}{GO term corresponding to the gene. Data type: character}
}
}
\usage{
data(GO_Term.example)
}
\description{
A dataset containing maize GO Terms of B73 version4 from agrigo(http://bioinfo.cau.edu.cn/agriGO/index.php)
}
\keyword{datasets}
