% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/function.Gprime.R
\name{Gprime}
\alias{Gprime}
\title{The function to perform BSA analysis with G-test algorithm}
\usage{
Gprime(file, Windowsize, threshold = 0.95, pval = 0.05)
}
\arguments{
\item{file}{The data frame exported by BSA_filter()}

\item{Windowsize}{A numeric value indicates the window size (in base pairs) containing SNPs subjected to calculating G statistic}

\item{threshold}{From 0 to 1, default value is “0.95”. A numeric threshold value to extract candidate region by fractile quantile}

\item{pval}{A numeric threshold value, default value is “0.05”}
}
\value{
a list containing three data frame,the data frame named "total.result" containing all the calculate result by Gtest algorithm,the data frame named "sigthreshold.result" containing the significant SNPs under fractile quantile set by parameter "threshold" calculated by Gtest algorithm,the data frame named "sigpval.result" containing the significant SNPs selected by P-value set by parameter "pval" calculated by Gtest algorithm
}
\description{
The function to perform BSA analysis with G-test algorithm
}
