% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gtf_data.example.R
\docType{data}
\name{gtf_data.example}
\alias{gtf_data.example}
\title{Genotype data}
\format{
gtf in GRanges format with 6 ranges and 15 metadata columns
\describe{
\item{seqnames}{The ID of the sequence, which can be the ID of chromosome, scaffold or contig.. Data type: factor}
\item{start}{The start position of the feature on the sequence. Data type: integer}
\item{end}{The end position of the feature on the sequence. Data type: integer}
\item{width}{The width of the feature on the sequence. Data type: integer}
\item{source}{The source of this file. Data type: factor}
\item{type}{The feature represented by the region between start and end can be gene, non coding RNA, TE, exon, trascript, etc. Data type: factor}
\item{score}{A floating-point value. When there is a value, it represents the reliability of the above features. Data type: numeric}
\item{phase}{Only valid for comment type "CDS", indicating the position of the starting code. The valid values are 0, 1 and 2. Data type: integer}
\item{gene_id}{The name of this sequence. Data type: character}
\item{gene_biotype}{optional column.The biological type of this sequence such as gene, lincRNA, snoRNA, etc. Data type: character}
\item{transcript_id}{The name of this sequence, if the sequence is a transcript. Data type: character}
\item{transcript_source}{optional column.The source of this transcript. Data type: character}
\item{transcript_biotype}{optional column.The biological type of this sequence. Data type: character}
\item{exon_number}{optional column.The exon number of the transcript. Data type: character}
\item{exon_id}{optional column.The name of exon if above feature is an exon. Data type: character}
\item{protein_id}{optional column.The name of corresponding protein. Data type: character}
\item{protein_version}{optional column.The version of protein function annotation. Data type: character}
\item{gene_name}{optional column.indicate the function of gene. Data type: character}
}
}
\usage{
data(gtf_data.example)
}
\description{
A dataset containing part of maize genes positional information of B73 version4 from ensembl plant(ftp://ftp.ensemblgenomes.org/pub/plants/release-34/gtf/zea_mays/Zea_mays.AGPv4.34.gtf.gz)
}
\keyword{datasets}
