% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/function.plot.R
\name{plot_BSA}
\alias{plot_BSA}
\title{The function to output BSA analysis result into scatter and line plots}
\usage{
plot_BSA(
  file,
  chromlist = NULL,
  type = c("deltaSNP", "Bayesian", "ED", "Gtest"),
  file.name,
  threshold = NULL,
  pval = NULL
)
}
\arguments{
\item{file}{The data frame exported by the BSA analysis functions}

\item{chromlist}{A string vector of the chromosome names used to plot line plots specific chromosomes, default value is "NULL"}

\item{type}{A character to indicate the selected algorithm}

\item{file.name}{A character to name the output plot}

\item{threshold}{From 0 to 1, default value is "NULL". The red threshold line displayed in the output plot}

\item{pval}{Default value is "NULL". A numeric threshold value drawn in blue line in the output plot}
}
\value{
a scatter plot for statistic value and a line plot for tricube-smoothed statistic value of whole genome, if parameter "chromlist" is not "NULL", will also output a series line plot for tricube-smoothed statistic value of each chromosome
}
\description{
The function to output BSA analysis result into scatter and line plots
}
