% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/function.resultanno.R
\name{result_anno}
\alias{result_anno}
\title{The function to extract genes fall in the candidate region and judge whether gene is differential expression}
\usage{
result_anno(BSA_data, gtf_data, DEG_data = NULL)
}
\arguments{
\item{BSA_data}{A data frame contains three columns, including the chromosome name, the start and the end positions of the candidate regions}

\item{gtf_data}{The name of reference genome in GTF format}

\item{DEG_data}{A vector containing the names of differential expression genes,default is "NULL"}
}
\value{
if parameter "DEG_data" is set "NULL" will return a data frame contain the genes fall in the candidate region,if paramater "DEG_data" is a vector provide the gene id of differential expression will return a list, the first data frame in the list contains the genes fall in the candidate region,the second data frame contains the differential expression genes fall in the candidate region
}
\description{
The function to extract genes fall in the candidate region and judge whether gene is differential expression
}
