% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/function.import.R
\name{vcf2table}
\alias{vcf2table}
\title{The function converts file in VCF format from GATK into a data frame}
\usage{
vcf2table(file, BulkA, BulkB, chromlist = NULL)
}
\arguments{
\item{file}{The name of genotype file in VCF format from GATK SNP calling pipeline}

\item{BulkA}{The sample name of one bulk}

\item{BulkB}{The sample name of the other bulk}

\item{chromlist}{Default value is “NULL”.A string vector of the chromosomes to be used in the analysis, used to filter out unneeded contigs etc}
}
\value{
a data frame containing columns for Read depth (DP),Reference Allele Depth (AD_REF), Alternative Allele Depth (AD_ALT) and Genoytype Quality (GQ) for each bulk (indicated by .Bulk1 and .Bulk2 column name suffix)
}
\description{
The function converts file in VCF format from GATK into a data frame
}
